+++ serde-encrypt/src/traits.rs
@@ -2,6 +2,8 @@
 
 mod serde_encrypt_public_key;
 mod serde_encrypt_shared_key;
+mod serde_encrypt_shared_key_deterministic;
 
 pub use serde_encrypt_public_key::SerdeEncryptPublicKey;
 pub use serde_encrypt_shared_key::SerdeEncryptSharedKey;
+pub use serde_encrypt_shared_key_deterministic::SerdeEncryptSharedKeyDeterministic;